requirements.txt
anyio==3.5.0
asgiref==3.5.0
async-generator==1.10
attrs==21.2.0
awscli==1.24.10
bcrypt==3.2.0
beautifulsoup4==4.10.0
boto3==1.23.9
botocore==1.26.10
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.7
click==8.0.3
colorama==0.4.4
configparser==5.2.0
cryptography==35.0.0
docutils==0.16
fastapi==0.73.0
Flask==2.1.1
greenlet==1.1.2
h11==0.12.0
idna==3.3
importlib-metadata==4.11.3
inflect==5.4.0
itsdangerous==2.1.2
Jinja2==3.1.1
jmespath==1.0.0
lxml==4.7.1
MarkupSafe==2.1.1
passlib==1.7.4
pyasn1==0.4.8
pycparser==2.20
pydantic==1.9.0
PyJWT==2.3.0
PyMySQL==1.0.2
python-dateutil==2.8.2
python-slugify==6.1.2
PyYAML==5.4.1
rsa==4.7.2
s3transfer==0.5.2
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.2.post1
SQLAlchemy==1.4.36
starlette==0.17.1
text-unidecode==1.3
typing-extensions==4.1.1
urllib3==1.26.7
uvicorn==0.17.6
Werkzeug==2.1.1
zipp==3.8.0
